+++ electron-svelte-ToDo-App/src/TodoItems.svelte
@@ -10,9 +10,13 @@
     const dispatch = createEventDispatcher();
 
     function deleteTodo(){
-        dispatch('deleteTodo', {
-            id: id
-        });
+        if (confirm('Are you sure you want to delet this entry?')) {
+            dispatch('deleteTodo', {
+                id: id 
+            });
+        } else {
+        // Do nothing!
+        }
     }
 
     function toggleComplete(){
